editor + MazeRunner
NEW:
When a button is selected with the left mouse button, it is surrounded
by a blue box. When a button is selected usng the right mouse button, it
is surrounded with a green box.

FIXED:
The maze built up was wrong, x and z where reversed.
the spikes are now occupying the whole square